+++ image_scripts/test/test-files/etc_dojo.d/scripts/29-not-executable-file.sh
@@ -1,3 +1,3 @@
 #!/bin/bash
 
-echo "Running a script which user forgot to make executable"
+echo "Running a script which user forgot to make executable" >&2